Switch to using Context instead of global Object variables (#61)
This is another approach to fixing #53 

This switches from using ObjectNames and ObjectValues to using a new set
of global dictionaries. ContextNames and ContextValues the primary
reason for the name change is because I'm trying not to change every
template at once. Eventually the "Object*" versions would be removed
when nothing is using them anymore.

The big difference here is that the new "Context*" are dicts. This way
multiple can be used without conflicting. Now when configmaps is
included the values are set on the "configmap" key of these dicts so it
doesn't conflict with the "app" in use by the templates that use
application context.

The other thing this did for me just as a template author is it made it
more clear and explicit what context each template expected and used.
For example Deployments now uses the "app" context b/c that's what it
expects to have and that's now explicit.

I've also added some checks in _all that the "Context*" dicts are empty
at the end of rendering. The idea being when you set a context, you
unset it when you're done with it. This helps prevent unexpectedly using
stale values. You'll see this done in a few places.

Additionally I modified names.fullname so that you can tell it which
context to build names from, the configmap template does this.

---------

Co-authored-by: Diamon Wiggins <diamonw757@gmail.com>